+++ src/Field.cpp
@@ -2,7 +2,7 @@
 
 #define DIVSQRT2 0.70710678118
 
-Field::Field(int rows, int cols) : fields(rows, cols, false), Rows(rows), Cols(cols)
+Field::Field(size_t rows, size_t cols) : fields(rows, cols, false), Rows(rows), Cols(cols)
 {
 }
